rename sc-bilara-segmented-text-styles to sc-layout-bilara-styles for consistency
